L21 - Implementing message publisher in Order Application Service to fire the events
Implemented event publishing.
